# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Neema Gospel Choir",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get list of donations",
                "operationId": "efb11d52c97c83578916f73e51b55f1f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of donations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Create a new donation",
                "operationId": "c40dee2dafc26f2f389fd4547dcade3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "campaign_id",
                                    "donor_name",
                                    "donor_email",
                                    "amount",
                                    "transaction_reference"
                                ],
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "donor_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "donor_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "donor_phone": {
                                        "type": "string",
                                        "example": "+255712345678"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 100
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "TZS"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "M-Pesa"
                                    },
                                    "transaction_reference": {
                                        "type": "string",
                                        "example": "TXN123456789"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "God bless you"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donations/{donation}": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get donation details",
                "operationId": "12b4d8419a234c193fc7990fbf43b038",
                "parameters": [
                    {
                        "name": "donation",
                        "in": "path",
                        "description": "Donation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Donations"
                ],
                "summary": "Update donation status",
                "operationId": "e4aaf5c1822f3869a341c38589b51709",
                "parameters": [
                    {
                        "name": "donation",
                        "in": "path",
                        "description": "Donation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "failed",
                                            "refunded"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Donations"
                ],
                "summary": "Delete a donation",
                "operationId": "c4654e7c6000ac9e3150274496935388",
                "parameters": [
                    {
                        "name": "donation",
                        "in": "path",
                        "description": "Donation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donations/campaign/{campaign}": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get donations by campaign",
                "operationId": "2fc506416cb9b48bcada73b055b190ca",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of donations for the campaign",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donations/statistics": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get donation statistics",
                "operationId": "e036b3deb0b71f0d0a3a8794b4822f4e",
                "responses": {
                    "200": {
                        "description": "Donation statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_donations": {
                                                    "type": "number",
                                                    "example": 15000
                                                },
                                                "total_donors": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "total_active_campaigns": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "today_donations": {
                                                    "type": "number",
                                                    "example": 500
                                                },
                                                "monthly_donations": {
                                                    "type": "number",
                                                    "example": 2500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}